Added combat-logic.js and combat-abilities.js
Created new files and classes to handle how combat is resolved. Combat-abilities contains the Skills class which will be a list of all skills along with any information required to handle their execution in game. Combat logic has the CombatLogic class which will handle the actual execution of combat abilities.