Created an on-screen text box to relay messages to player.
Creatred an on-screen text box to relay information to the player during combat. The text box is positioned beneath the Combat Field, taking over the space the row of skulls once held.
Slowly adding useful information such as: when a unit cannot act, explainign why, displaying damage done and health healed when a skill is used, etc. This will likely be an ongoing process as different pieces of information become pertinent to the player.