Creature skills now based on their component pieces
From the game's inception, the idea was that the Creatures controlle by the player would be undead constructs (think Frankenstein's Monster), built with pieces found here and there. The intent was that the parts you use to build your creatures determine what abilities they have in combat, as well as affecting their stats. As of today, a Creature's guts, skin, and bones determine their skills.